teams corrections has done as per the PR suggestions ,'controller splited to two controllers ,usage of verbs and action words has been changed'
