+++ app/views/dashboards/_index_team_desktop.html.erb
@@ -1,5 +1,4 @@
 <%= render "shared/components/desktop_layout" do %>
-<main class="" >
    <div data-controller="teams">
       <h1 class="mb-4 pl-0 text-xl font-bold">My Team</h1>
       <div class="flex justify-between">
@@ -26,5 +25,4 @@
       <%= render "shared/components/team_list" %> 
       <%= render "shared/components/team_add_component" %>
    </div>
-</main>
 <% end %>
