+++ app/views/shared/components/_team_user_info.html.erb
@@ -1,22 +1,22 @@
-<div class="p-6 bg-white border rounded box-shadow-medium mt-6" data-controller="team-user-info">
+<div class="p-6 bg-white border rounded box-shadow-medium mt-6" data-controller="team-user-status user-status-confirmation">
     <div class="flex justify-between pb-4">
        <p class="text-xl font-semibold text-primary">
           Balaji Balachandran
        </p>
        <div class="flex space-x-4">
-          <div class="flex items-center space-x-2 bg-primary-light-50 py-1 px-2 rounded" data-team-user-info-target="active">
+          <div class="flex items-center space-x-2 bg-primary-light-50 py-1 px-2 rounded" data-user-status-confirmation-target="activeUser">
              <span class="bg-cover inline-block icon-green-circle  h-2 w-2"></span>
              <p class="text-sm text-primary">Active</p>
           </div>
-          <div class="flex items-center space-x-2 bg-danger-light py-1 px-2 rounded hidden" data-team-user-info-target="deActive">
+          <div class="flex items-center space-x-2 bg-danger-light py-1 px-2 rounded hidden" data-user-status-confirmation-target="inactiveUser">
             <span class="bg-cover inline-block icon-red-circle  h-2 w-2"></span>
-            <p class="text-sm text-danger">Deactive</p>
+            <p class="text-sm text-danger">Deactivated</p>
          </div>
-          <label for="" class="flex items-center cursor-pointer" data-action="click->team-user-info#toggleDropdown">
+          <label for="" class="flex items-center cursor-pointer" data-action="click->team-user-status#toggleDropdown">
           <span class="h-4 w-4 icon icon-more-vertical"></span>
           </label> 
-          <div class="absolute right-6 mt-6 bg-white border rounded shadow-md hidden deactivate-menu">
-            <p class="text-xm cursor-pointer px-2 py-3" data-action="click->team-user-info#showConfirmation" data-team-user-info-target="deActiveMenu">Deactivate</p>
+          <div class="absolute right-6 mt-6 bg-white border rounded shadow-md hidden" data-team-user-status-target="deActivateMenu" data-user-status-confirmation-target="deActivateMenu">
+            <p class="text-xm cursor-pointer px-2 py-3" data-action="click->team-user-status#showConfirmation">Deactivate</p>
         </div>
        </div>
     </div>
