+++ app/views/shared/components/_team_user_status_popup.html.erb
@@ -1,28 +1,34 @@
-<div class="fixed inset-0 bg-black bg-opacity-75 hidden deactivate-confirmation">
+<div class="fixed inset-0 bg-black bg-opacity-75 hidden" data-team-user-status-target="deactivateConfirmation" data-user-status-confirmation-target="deactivateConfirmation">
     <div class="flex items-center justify-center min-h-screen">
         <div class="bg-white rounded shadow-md">
-            <div class="border-b">
-                <h2 class="text-base font-semibold p-6">Deactivate</h2>
+            <div class="border-b flex justify-between p-6">
+                <h2 class="text-base font-semibold">Deactivate</h2>
+                <button id="cancel-btn-deactive" type="button" data-action="click->user-status-confirmation#cancel">
+                    <span class="icon icon-close h-4 w-4"></span>
+                    </button>
             </div>
             <p class="text-sm mb-4 p-6">Do you want Deactivate <span>Balaji Balachandran</span> ?</p>
             <div class="px-6 pb-6 flex flex-col">
-                <button data-action="click->team-user-info#confirm" id="de-activate-yes" class="bg-danger text-white px-4 py-2 rounded w-[414px] h-14">Yes</button>
-                <button data-action="click->team-user-info#cancel" id="de-activate-no" class="  px-4 py-2 w-[414px] h-14">No</button>
+                <button data-action="click->user-status-confirmation#confirm" id="de-activate-yes" class="bg-danger text-white px-4 py-2 rounded w-[414px] h-14">Yes</button>
+                <button data-action="click->user-status-confirmation#cancel" id="de-activate-no" class="  px-4 py-2 w-[414px] h-14">No</button>
             </div>
         </div>
     </div>
 </div>
 
-<div class="fixed inset-0 bg-black bg-opacity-75 hidden activate-confirmation">
+<div class="fixed inset-0 bg-black bg-opacity-75 hidden" data-team-user-status-target="activateConfirmation" data-user-status-confirmation-target="activateConfirmation">
   <div class="flex items-center justify-center min-h-screen">
       <div class="bg-white rounded shadow-md">
-          <div class="border-b">
-              <h2 class="text-base font-semibold p-6">Activate</h2>
+          <div class="border-b flex justify-between p-6">
+              <h2 class="text-base font-semibold ">Activate</h2>
+              <button id="cancel-btn-active" type="button" data-action="click->user-status-confirmation#cancel">
+                <span class="icon icon-close h-4 w-4"></span>
+                </button>
           </div>
           <p class="text-sm mb-4 p-6">Do you want Activate <span>Balaji Balachandran</span> ?</p>
           <div class="px-6 pb-6 flex flex-col">
-              <button data-action="click->team-user-info#confirm" id="activate-yes" class="bg-primary text-white px-4 py-2 rounded w-[414px] h-14">Yes</button>
-              <button data-action="click->team-user-info#cancel" id="activate-no" class="  px-4 py-2 w-[414px] h-14">No</button>
+              <button data-action="click->user-status-confirmation#confirm" id="activate-yes" class="bg-primary text-white px-4 py-2 rounded w-[414px] h-14">Yes</button>
+              <button data-action="click->user-status-confirmation#cancel" id="activate-no" class="  px-4 py-2 w-[414px] h-14">No</button>
           </div>
       </div>
   </div>
